requirements.txt
elevenlabs==0.2.27
flask==3.1.1
google-generativeai==0.8.5
python-dotenv==1.1.1